+++ ecc/bls24-315/fr/element.go
@@ -20,7 +20,7 @@ import (
 	"crypto/rand"
 	"encoding/binary"
 	"errors"
-	"github.com/consensys/gnark-crypto/internal/hashutils"
+	"github.com/consensys/gnark-crypto/field"
 	"io"
 	"math/big"
 	"math/bits"
@@ -768,7 +768,7 @@ func Hash(msg, dst []byte, count int) ([]Element, error) {
 	const L = 16 + Bytes
 
 	lenInBytes := count * L
-	pseudoRandomBytes, err := hashutils.ExpandMsgXmd(msg, dst, lenInBytes)
+	pseudoRandomBytes, err := field.ExpandMsgXmd(msg, dst, lenInBytes)
 	if err != nil {
 		return nil, err
 	}
