+++ internal/field/internal/templates/element/tests.go
@@ -10,7 +10,7 @@ import (
 	"math/bits"
 	"fmt"
 	{{if .UsingP20Inverse}} 
-	"github.com/consensys/gnark-crypto/internal/field"
+	"github.com/consensys/gnark-crypto/field"
 	mrand "math/rand" 
 	{{end}}
 	"testing"
